refactor : Session 패키지로 이동
