+++ src/main/java/nextstep/courses/domain/session/SessionRepository.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 public interface SessionRepository {
